Extract command to interface
